ScareBnB-Backend/requirements.txt
aioboto3==9.6.0
aiobotocore==2.3.0
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
blinker==1.4
boto3==1.21.21
botocore==1.24.21
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
executing==0.8.3
filestorage==0.0.6
flake8==4.0.1
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
ipython==8.4.0
is-disposable-email==1.0.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
multidict==6.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
stack-data==0.3.0
toml==0.10.2
traitlets==5.3.0
urllib3==1.26.11
uuid
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
wtforms-validators==1.0.0
yarl==1.7.2
